wip: Complete structure refactor
- Endpoint models removed and rafactored into 'Responses'
- Responses use base classes
- ClimaCellService 1/4 responses implemented

+++ src/ClimaCellCore/Models/Layers/AirQuality.cs
@@ -5,7 +5,7 @@
     public class Pm25
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -14,7 +14,7 @@
     public class Pm10
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -23,7 +23,7 @@
     public class O3
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -32,7 +32,7 @@
     public class NO2
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -41,7 +41,7 @@
     public class CO
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -50,7 +50,7 @@
     public class SO2
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -59,7 +59,7 @@
     public class EpaAQI
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
     }
 
     public class EpaPrimaryPollutant
@@ -77,7 +77,7 @@
     public class ChinaAQI
     {
         [JsonProperty("value")]
-        public float Value { get; set; }
+        public double Value { get; set; }
     }
 
     public class ChinaPrimaryPollutant
