+++ src/ClimaCellCore/Models/Layers/Pollen.cs
@@ -5,7 +5,7 @@
     public class PollenTree
     {
         [JsonProperty("value")]
-        public int Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -14,7 +14,7 @@
     public class PollenWeed
     {
         [JsonProperty("value")]
-        public int Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
@@ -23,7 +23,7 @@
     public class PollenGrass
     {
         [JsonProperty("value")]
-        public int Value { get; set; }
+        public double Value { get; set; }
 
         [JsonProperty("units")]
         public string Units { get; set; }
